make actions/scripts top-level; move utils
