+++ .github/workflows/retagger.yml
@@ -1,0 +1,15 @@
+on:
+  release:
+    types:
+      - created
+      - edited
+permissions:
+  contents: write
+jobs:
+  retag:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: fischerscode/tagger@v0
+        with:
+          prefix: v
