+++ actions/get-platform/action.yml
@@ -28,7 +28,7 @@ runs:
       with:
         srcroot: null
 
-    - uses: teaxyz/brewkit/.github/actions/cache@main
+    - uses: teaxyz/brewkit/actions/cache@v0
       with:
         cache-name: get-platform
