+++ actions/get-platform/get-platform.ts
