+++ actions/setup-brewkit/action.yml
@@ -28,7 +28,7 @@ runs:
         # prevent pantry from reassigning TEA_PREFIX etc.
         srcroot: null
 
-    - uses: teaxyz/brewkit/.github/actions/cache@main
+    - uses: teaxyz/brewkit/actions/cache@v0
       with:
         cache-name: setup
