+++ actions/stage-build-artifacts/action.yml
@@ -24,7 +24,7 @@ runs:
 
     - uses: teaxyz/setup@v0
 
-    - uses: teaxyz/brewkit/.github/actions/cache@main
+    - uses: teaxyz/brewkit/actions/cache@v0
       with:
         cache-name: stage
