+++ actions/upload/action.yml
@@ -34,7 +34,7 @@ outputs:
 runs:
   using: composite
   steps:
-    - uses: teaxyz/brewkit/.github/actions/cache@main
+    - uses: teaxyz/brewkit/actions/cache@v0
       with:
         cache-name: upload
